Sam: refactor recipes page to user master-detail pattern
