--- lib/sam_layouts/recipe_listing.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:swiftcook/model/recipe.dart';
-
-class RecipeListing extends StatelessWidget {
-  const RecipeListing(
-      {@required this.recipeSelectedCallback, this.selectedRecipe});
-
-  final ValueChanged<Recipe> recipeSelectedCallback;
-  final Recipe selectedRecipe;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView(
-        children: ListTile.divideTiles(
-      context: context,
-      tiles: Recipe.getTestData().map((recipe) {
-        // in the future this should probably be generated by some backend code
-        return ListTile(
-            title: Text(recipe.title),
-            onTap: () => recipeSelectedCallback(recipe),
-            selected: selectedRecipe == recipe);
-      }).toList(),
-    ).toList());
-  }
-}
